The start of a system to use to play cards?
The idea here will be to create a few different classes (ranged attack, melee attack, buff, debuff, etc) and each of them will have a different base PlayCard implementation so that I can easily add more. Not sure if this will pan out but some of it seems to make sense?